Created package predicate in module. Moved NameContainsKeywordsPredicate.java and SameModuleCodePredicate.java to predicate.
