+++ src/main/java/seedu/module/model/Model.java
@@ -8,7 +8,7 @@ import javafx.collections.ObservableList;
 import seedu.module.commons.core.GuiSettings;
 import seedu.module.model.module.ArchivedModule;
 import seedu.module.model.module.Module;
-import seedu.module.model.module.SameModuleCodePredicate;
+import seedu.module.model.module.predicate.SameModuleCodePredicate;
 import seedu.module.model.module.TrackedModule;
 
 /**
