+++ src/main/java/seedu/module/model/module/predicate/NameContainsKeywordsPredicate.java
@@ -1,9 +1,10 @@
-package seedu.module.model.module;
+package seedu.module.model.module.predicate;
 
 import java.util.List;
 import java.util.function.Predicate;
 
 import seedu.module.commons.util.StringUtil;
+import seedu.module.model.module.Module;
 
 /**
  * Tests that a {@code Module}'s moduleCode matches any of the keywords given.
