+++ src/test/java/seedu/module/logic/commands/AddCommandTest.java
@@ -14,7 +14,7 @@ import seedu.module.model.ModelManager;
 import seedu.module.model.ModuleBook;
 import seedu.module.model.module.ArchivedModule;
 import seedu.module.model.module.ArchivedModuleList;
-import seedu.module.model.module.SameModuleCodePredicate;
+import seedu.module.model.module.predicate.SameModuleCodePredicate;
 import seedu.module.model.module.TrackedModule;
 import seedu.module.testutil.ArchivedModuleBuilder;
