+++ src/test/java/seedu/module/logic/commands/DeleteCommandTest.java
@@ -11,7 +11,7 @@ import org.junit.jupiter.api.Test;
 import seedu.module.model.Model;
 import seedu.module.model.ModelManager;
 import seedu.module.model.ModuleBook;
-import seedu.module.model.module.SameModuleCodePredicate;
+import seedu.module.model.module.predicate.SameModuleCodePredicate;
 import seedu.module.model.module.TrackedModule;
 import seedu.module.testutil.TrackedModuleBuilder;
