+++ src/test/java/seedu/module/logic/parser/ModuleBookParserTest.java
@@ -22,8 +22,8 @@ import seedu.module.logic.commands.ListCommand;
 import seedu.module.logic.commands.ViewCommand;
 import seedu.module.logic.parser.exceptions.ParseException;
 import seedu.module.model.module.ArchivedModule;
-import seedu.module.model.module.NameContainsKeywordsPredicate;
-import seedu.module.model.module.SameModuleCodePredicate;
+import seedu.module.model.module.predicate.NameContainsKeywordsPredicate;
+import seedu.module.model.module.predicate.SameModuleCodePredicate;
 import seedu.module.model.module.TrackedModule;
 import seedu.module.testutil.ArchivedModuleBuilder;
 import seedu.module.testutil.TrackedModuleBuilder;
